Move clip state into vello_common
